REFACTOR upratovanie + TESTS PASSED
