+++ src/plugins/_shared/components/form/Details.tsx
