+++ src/plugins/app/footer/PrivacyPolicyLink.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { ExternalLink } from '../../_components/links/ExternalLink'
+import { ExternalLink } from '../../_shared/components/links/ExternalLink'
 
 export const PrivacyPolicyLink = () => (
   <ExternalLink href="https://navody.digital/ochrana-osobnych-udajov">
