+++ src/plugins/steps/1-prijmy/prijmy-a-vydavky.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { Form, FormikProps } from 'formik'
-import { FormWrapper, Input } from '../../_components/form/FormComponents'
+import { FormWrapper, Input } from '../../_shared/components/form/FormComponents'
 import { FormErrors, IncomeAndExpenseUserInput } from '../../_shared/types/PageUserInputs'
 import { numberInputRegexp } from '../../_shared/utils/utils'
-import { ErrorSummary } from '../../_components/form/ErrorSummary'
+import { ErrorSummary } from '../../_shared/components/form/ErrorSummary'
 import { Page } from '../../_shared/interfaces/Page'
-import { BackLink } from '../../_components/links/BackLink'
+import { BackLink } from '../../_shared/components/links/BackLink'
 import { TAX_YEAR } from '../../_shared/calculation/calculation'
 
 const PrijmyAVydavky: Page<IncomeAndExpenseUserInput> = ({
