+++ src/plugins/steps/3-manzelstvo/PartnerBonusFormSteps.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
-import { BooleanRadio, CheckboxSmall, Input } from '../../_components/form/FormComponents'
+import { BooleanRadio, CheckboxSmall, Input } from '../../_shared/components/form/FormComponents'
 import { formatRodneCislo } from '../../_shared/utils/utils'
 import { PartnerBonusFormProps } from './PartnerBonusForm'
-import { Details } from '../../_components/form/Details'
+import { Details } from '../../_shared/components/form/Details'
 import { TAX_YEAR } from '../../_shared/calculation/calculation'
 
 export const ApplyForBonusQuestion = ({ disabled }) => (
