+++ src/plugins/steps/3c-hypoteka/hypoteka.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import Link from 'next/link'
 import { Form } from 'formik'
-import { BooleanRadio, FormWrapper, Input } from '../../_components/form/FormComponents'
+import { BooleanRadio, FormWrapper, Input } from '../../_shared/components/form/FormComponents'
 import { FormErrors, MortgageUserInput } from '../../_shared/types/PageUserInputs'
-import { ErrorSummary } from '../../_components/form/ErrorSummary'
+import { ErrorSummary } from '../../_shared/components/form/ErrorSummary'
 import { numberInputRegexp } from '../../_shared/utils/utils'
 import { Page } from '../../_shared/interfaces/Page'
 import { mortgageInitialValues } from '../../_shared/calculation/initialValues'
