+++ src/plugins/steps/3d-dve-percenta/dve-percenta.tsx
@@ -6,19 +6,19 @@ import {
   CheckboxSmall,
   FormWrapper,
   Input,
-} from '../../_components/form/FormComponents'
+} from '../../_shared/components/form/FormComponents'
 import { FormErrors, TwoPercentUserInput } from '../../_shared/types/PageUserInputs'
 import styles from '../5-osobne-udaje/osobne-udaje.module.css'
 import { getNgoByName } from '../../api/_utils/api'
-import { ErrorSummary } from '../../_components/form/ErrorSummary'
+import { ErrorSummary } from '../../_shared/components/form/ErrorSummary'
 import {
   AutoCompleteData,
   AutoCompleteInput,
-} from '../../_components/form/AutoCompleteInput'
+} from '../../_shared/components/form/AutoCompleteInput'
 
 import { Page } from '../../_shared/interfaces/Page'
 import { twoPercentInitialValues } from '../../_shared/calculation/initialValues'
-import { Details } from '../../_components/form/Details'
+import { Details } from '../../_shared/components/form/Details'
 import classNames from 'classnames'
 
 const makeHandleOrganisationAutoform = ({
