+++ src/plugins/summary/8-vypocet-dane/vysledky.tsx
@@ -3,7 +3,7 @@ import { formatCurrency } from '../../_shared/utils/utils'
 import { TaxFormUserInput } from '../../_shared/types/TaxFormUserInput'
 import { Page } from '../../_shared/interfaces/Page'
 import Decimal from 'decimal.js'
-import { BackLink } from '../../_components/links/BackLink'
+import { BackLink } from '../../_shared/components/links/BackLink'
 import Link from 'next/link'
 import { buildSummary } from '../../_shared/calculation/calculation'
 import { TAX_YEAR } from '../../_shared/calculation/calculation'
