+++ src/plugins/summary/9-pokracovat/pokracovat.tsx
@@ -3,7 +3,7 @@ import getConfig from 'next/config'
 import { Page } from '../../_shared/interfaces/Page'
 import { TaxForm } from '../../_shared/types/TaxForm'
 import { convertToXML } from './xml/xmlConverter'
-import { RedirectField, RedirectForm } from '../../_components/form/RedirectForm'
+import { RedirectField, RedirectForm } from '../../_shared/components/form/RedirectForm'
 import { setDate, toBase64, formatCurrency } from '../../_shared/utils/utils'
 import { buildSummary } from '../../_shared/calculation/calculation'
 import { Summary } from './Summary.types'
